refactor(migrate): modularize openclaw flow and harden dry-run scan
